Renamed ProtoEval project to ProtoBufEval to be consistent with the package id
