+++ src/ProtoBufEval.Tests/Usings.cs
@@ -1,0 +1,2 @@
+global using Xunit;
+global using ProtoBufEval;
