+++ src/ProtoBufEval.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ProtoEval", "ProtoEval\ProtoEval.csproj", "{9AA21403-9D6F-4780-9119-F85FAAB78700}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ProtoBufEval", "ProtoBufEval\ProtoBufEval.csproj", "{9AA21403-9D6F-4780-9119-F85FAAB78700}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ProtoEval.Tests", "ProtoEval.Tests\ProtoEval.Tests.csproj", "{7E3690C1-D41E-465D-BCD5-015320CFC7A7}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ProtoBufEval.Tests", "ProtoBufEval.Tests\ProtoBufEval.Tests.csproj", "{7E3690C1-D41E-465D-BCD5-015320CFC7A7}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
